backend/requirements.txt
# Django & Core Web
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
daphne
channels
asgiref==3.9.1

# API & HTTP
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
urllib3==2.5.0
aiohttp==3.13.1
aiohttp_socks==0.10.1

# OpenAI
openai==1.99.9

# Google Cloud & APIs
google-cloud-storage
google-cloud-texttospeech==2.27.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0

# Pinecone (use pinecone, not pinecone-client)
pinecone==6.0.2
pinecone-plugin-interface==0.0.7

# PyTorch & Transformers (for SigLIP2)
torch==2.7.0
torchvision==0.22.0
transformers==4.57.1
tokenizers==0.22.1
accelerate==1.11.0
safetensors==0.6.2
huggingface-hub>=0.34.0,<1.0

# Image Processing
Pillow==11.3.0
opencv-python==4.12.0.88
scikit-image==0.25.2
imageio==2.37.0

# Audio Processing
pydub==0.25.1
mutagen==1.47.0

# Web Scraping & Research
beautifulsoup4==4.14.2
bs4==0.0.2
lxml==6.0.2
duckduckgo_search==8.1.1
tldextract==5.3.0
requests-html==0.10.0

# NLP
nltk==3.9.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
tiktoken==0.9.0

# Data Validation & Types
pydantic==2.11.7
pydantic_core==2.33.2
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.1

# Database
sqlparse==0.5.3
SQLAlchemy==2.0.44
psycopg==3.2.6
psycopg2-binary==2.9.10
alembic==1.17.0

# Utilities
python-dotenv==1.1.1
tqdm==4.67.1
colorama==0.4.6
PyJWT==2.10.1
PyYAML==6.0.3
Jinja2==3.1.6
MarkupSafe==3.0.3
regex==2025.10.23
filelock==3.20.0
packaging==25.0
click==8.3.0
rich==14.2.0

# Scientific Computing
numpy
scipy==1.16.3
scikit-learn==1.7.2
sympy==1.14.0
networkx==3.5

# Crypto & Auth
cryptography==46.0.3
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
certifi==2025.8.3

# Async & Concurrency
anyio==4.10.0
sniffio==1.3.1
aiosignal==1.4.0
frozenlist==1.8.0
greenlet==3.2.4

# Protobuf & Serialization
protobuf==6.31.1
proto-plus==1.26.1

# ML & Vision Libraries
einops==0.8.1
timm==1.0.20
kornia==0.8.1
kornia_rs==0.1.9
spandrel==0.4.1

# ComfyUI (optional - for image generation)
# comfyui-embedded-docs==0.3.0
# comfyui_frontend_package==1.28.7
# comfyui_workflow_templates==0.2.1

# Misc
cachetools==5.5.2
charset-normalizer==3.4.3
distro==1.9.0
httplib2==0.22.0
idna==3.10
jiter==0.10.0
pyparsing==3.2.3
uritemplate==4.2.0
tzdata==2025.2
six==1.17.0
attrs==25.4.0
joblib==1.5.2
fsspec==2025.9.0
